Begin refactoring/deduplication of code between extract/rest of project
